Reduce execution times for mock workflow (#144)
The mock workflow was implemented to simulate realistic execution times for a
simulation+reconstruction workflow.  This results in a lengthy unit test (~26 seconds)
whose algorithms "spin" for the configured duration.  For the purposes of framework
development, it is not necessary to use such long durations.  We can restore the original
configured durations when it is deemed necessary.